chore: switch from netlify to now for survey function
