+++ now.json
@@ -1,5 +1,10 @@
 {
   "version": 2,
   "alias": "carbondesignsystem",
-  "public": true
+  "public": true,
+  "build": {
+    "env": {
+      "NODEJS_AWS_HANDLER_NAME": "survey"
+    }
+  }
 }
