+++ src/gatsby-theme-carbon/components/FeedbackDialog/FeedbackDialog.js
@@ -3,7 +3,7 @@ import ThemeFeedbackDialog from 'gatsby-theme-carbon/src/components/FeedbackDial
 
 const FeedbackDialog = ({ props }) => {
   const onSubmit = async (data) => {
-    const res = await fetch('/api/survey', {
+    const res = await fetch('https://carbondesignsystem.now.sh/api/survey', {
       method: 'POST',
       body: JSON.stringify(data),
       credentials: 'include',
